axfs_ramfs

Struct DirNode

pub struct DirNode { /* private fields */ }

The directory node in RAM filesystem.

This represents a directory that can contain files and subdirectories. It implements the VFS node operations trait to provide directory operations.

Fields

this - Weak reference to self for creating child directories

parent - Weak reference to parent directory

children - Map of child node names to their references

Implementations

impl DirNode

pub fn get_entries(&self) -> Vec<String>

Returns a list of all entry names in this directory.

Returns

A vector containing the names of all child nodes.

pub fn exist(&self, name: &str) -> bool

Checks whether a node with the given name exists in this directory.

Arguments

name - The name to check for

Returns

true if a node with the given name exists, false otherwise.

pub fn create_node(&self, name: &str, ty: VfsNodeType) -> VfsResult

Creates a new node with the given name and type in this directory.

Arguments

name - The name for the new node

ty - The type of node to create (file or directory)

Returns

Returns Ok(()) if the node was created, or an error if creation failed.

Errors

Returns VfsError::AlreadyExists if a node with the same name exists. Returns VfsError::Unsupported if the node type is not supported.

pub fn remove_node(&self, name: &str) -> VfsResult

Removes a node by the given name in this directory.

Directories can only be removed if they are empty.

Arguments

name - The name of the node to remove

Returns

Returns Ok(()) if the node was removed, or an error otherwise.

Errors

Returns VfsError::NotFound if the node does not exist. Returns VfsError::DirectoryNotEmpty if attempting to remove a non-empty directory.